requirements.txt
Flask==3.0.3
pypdf==5.1.0
python-docx==1.1.2
playwright==1.44.0
Jinja2==3.1.4
python-dotenv==1.0.1
gunicorn==22.0.1
